+++ errors/types_generated.go
@@ -15,8 +15,9 @@ var (
 	BetaInsufficientPermissions = Definitions[10]
 	BackendOpenapiValidation    = Definitions[12]
 	BackendTimeout              = Definitions[13]
-	Sequence                    = Definitions[15]
-	UnexpectedStatus            = Definitions[16]
+	BackendUnhealthy            = Definitions[14]
+	Sequence                    = Definitions[16]
+	UnexpectedStatus            = Definitions[17]
 )
 
 // typeDefinitions holds all related error definitions which are
@@ -40,6 +41,7 @@ var types = typeDefinitions{
 	"backend":                        Backend,
 	"backend_openapi_validation":     BackendOpenapiValidation,
 	"backend_timeout":                BackendTimeout,
+	"backend_unhealthy":              BackendUnhealthy,
 	"endpoint":                       Endpoint,
 	"sequence":                       Sequence,
 	"unexpected_status":              UnexpectedStatus,
@@ -54,4 +56,4 @@ func IsKnown(errorType string) bool {
 
 // SuperTypesMapsByContext holds maps for error super-types to sub-types
 // by a given context block type (e.g. api or endpoint).
-var SuperTypesMapsByContext = map[string]map[string][]string{"api": map[string][]string{"*": []string{"beta_insufficient_permissions", "backend_openapi_validation", "backend_timeout"}, "access_control": []string{"beta_insufficient_permissions"}, "backend": []string{"backend_openapi_validation", "backend_timeout"}}, "endpoint": map[string][]string{"*": []string{"beta_insufficient_permissions", "backend_openapi_validation", "backend_timeout", "sequence", "unexpected_status"}, "access_control": []string{"beta_insufficient_permissions"}, "backend": []string{"backend_openapi_validation", "backend_timeout"}, "endpoint": []string{"sequence", "unexpected_status"}}}
+var SuperTypesMapsByContext = map[string]map[string][]string{"api": map[string][]string{"*": []string{"beta_insufficient_permissions", "backend_openapi_validation", "backend_timeout", "backend_unhealthy"}, "access_control": []string{"beta_insufficient_permissions"}, "backend": []string{"backend_openapi_validation", "backend_timeout", "backend_unhealthy"}}, "endpoint": map[string][]string{"*": []string{"beta_insufficient_permissions", "backend_openapi_validation", "backend_timeout", "backend_unhealthy", "sequence", "unexpected_status"}, "access_control": []string{"beta_insufficient_permissions"}, "backend": []string{"backend_openapi_validation", "backend_timeout", "backend_unhealthy"}, "endpoint": []string{"sequence", "unexpected_status"}}}